pyproject.toml
[project]
name = "mlh-pe-hackathon"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "faker>=33.0",
    "fastapi>=0.135.3",
    "flask>=3.1",
    "orjson>=3.11.8",
    "peewee>=3.17",
    "psutil>=6.1.0",
    "psycopg2-binary>=2.9",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.22",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.43.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
]

[tool.pytest.ini_options]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=50"
testpaths = ["tests"]
